refactor: access token refresh; fix: search page error when refreshing without access token
